# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ConsultaMed API",
    "description": "API para gestión de consultas médicas",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/patients/": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "List Patients",
        "description": "List patients with optional search.\n\n- Search by partial name or DNI (minimum 2 characters)\n- Paginated results",
        "operationId": "list_patients_api_v1_patients__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by name or DNI",
              "title": "Search"
            },
            "description": "Search by name or DNI"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Create Patient",
        "description": "Create new patient.\n\nValidates:\n- DNI format (Spanish DNI/NIE)\n- DNI uniqueness\n- Required fields",
        "operationId": "create_patient_api_v1_patients__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Get Patient",
        "description": "Get patient by ID.\n\nReturns full patient data including allergies and recent encounters.",
        "operationId": "get_patient_api_v1_patients__patient_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Patients"
        ],
        "summary": "Update Patient",
        "description": "Update patient data.",
        "operationId": "update_patient_api_v1_patients__patient_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/allergies": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "List Allergies",
        "description": "List patient allergies.",
        "operationId": "list_allergies_api_v1_patients__patient_id__allergies_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AllergyResponse"
                  },
                  "title": "Response List Allergies Api V1 Patients  Patient Id  Allergies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Add Allergy",
        "description": "Add allergy to patient.",
        "operationId": "add_allergy_api_v1_patients__patient_id__allergies_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllergyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllergyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patients/{patient_id}/allergies/{allergy_id}": {
      "delete": {
        "tags": [
          "Patients"
        ],
        "summary": "Remove Allergy",
        "description": "Remove allergy from patient.",
        "operationId": "remove_allergy_api_v1_patients__patient_id__allergies__allergy_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "allergy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Allergy Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/encounters/patient/{patient_id}": {
      "get": {
        "tags": [
          "Encounters"
        ],
        "summary": "List Patient Encounters",
        "description": "List all encounters for a patient.\n\nReturns encounters with conditions and medications.",
        "operationId": "list_patient_encounters_api_v1_encounters_patient__patient_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Encounters"
        ],
        "summary": "Create Encounter",
        "description": "Create new encounter for patient.\n\nCreates:\n- Encounter record\n- Condition(s) for diagnoses\n- MedicationRequest(s) for prescriptions",
        "operationId": "create_encounter_api_v1_encounters_patient__patient_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncounterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllergyCreate": {
        "properties": {
          "code_text": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Code Text"
          },
          "type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(allergy|intolerance)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(food|medication|environment|biologic)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "criticality": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(low|high|unable-to-assess)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criticality"
          }
        },
        "type": "object",
        "required": [
          "code_text"
        ],
        "title": "AllergyCreate",
        "description": "Schema for creating an allergy."
      },
      "AllergyResponse": {
        "properties": {
          "code_text": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Code Text"
          },
          "type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(allergy|intolerance)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(food|medication|environment|biologic)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "criticality": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(low|high|unable-to-assess)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criticality"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "clinical_status": {
            "type": "string",
            "title": "Clinical Status"
          },
          "recorded_date": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded Date"
          }
        },
        "type": "object",
        "required": [
          "code_text",
          "id",
          "clinical_status",
          "recorded_date"
        ],
        "title": "AllergyResponse",
        "description": "Schema for allergy response."
      },
...
      "ConditionCreate": {
        "properties": {
          "code_text": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Code Text"
          },
          "code_coding_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Coding Code"
          }
        },
        "type": "object",
        "required": [
          "code_text"
        ],
        "title": "ConditionCreate",
        "description": "Schema for creating a condition."
      },
...
      "EncounterCreate": {
        "properties": {
          "reason_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Text"
          },
          "subjective_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjective Text"
          },
          "objective_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective Text"
          },
          "assessment_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Text"
          },
          "plan_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Text"
          },
          "recommendations_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendations Text"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionCreate"
            },
            "type": "array",
            "title": "Conditions",
            "default": []
          },
          "medications": {
            "items": {
              "$ref": "#/components/schemas/MedicationCreate"
            },
            "type": "array",
            "title": "Medications",
            "default": []
          }
        },
        "type": "object",
        "title": "EncounterCreate",
        "description": "Schema for creating an encounter."
      },
      "EncounterListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EncounterResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "EncounterListResponse",
        "description": "Paginated encounter list."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientCreate": {
        "properties": {
          "identifier_value": {
            "type": "string",
            "maxLength": 9,
            "minLength": 9,
            "title": "Identifier Value",
            "description": "DNI/NIE"
          },
          "name_given": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Given"
          },
          "name_family": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Family"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(male|female|other|unknown)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "telecom_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telecom Phone"
          },
          "telecom_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telecom Email"
          }
        },
        "type": "object",
        "required": [
          "identifier_value",
          "name_given",
          "name_family",
          "birth_date"
        ],
        "title": "PatientCreate",
        "description": "Schema for creating a patient."
      },
      "PatientListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatientSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PatientListResponse",
        "description": "Paginated list of patients."
      },
      "PatientResponse": {
        "properties": {
          "identifier_value": {
            "type": "string",
            "maxLength": 9,
            "minLength": 9,
            "title": "Identifier Value",
            "description": "DNI/NIE"
          },
          "name_given": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Given"
          },
          "name_family": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name Family"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(male|female|other|unknown)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "telecom_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telecom Phone"
          },
          "telecom_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telecom Email"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "allergies": {
            "items": {
              "$ref": "#/components/schemas/AllergyResponse"
            },
            "type": "array",
            "title": "Allergies",
            "default": []
          },
          "meta_created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Meta Created At"
          },
          "meta_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Meta Updated At"
          }
        },
        "type": "object",
        "required": [
          "identifier_value",
          "name_given",
          "name_family",
          "birth_date",
          "id",
          "age",
          "meta_created_at",
          "meta_updated_at"
        ],
        "title": "PatientResponse",
        "description": "Full patient response with allergies and encounters."
      },
...
      "PatientUpdate": {
        "properties": {
          "name_given": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Given"
          },
          "name_family": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Family"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(male|female|other|unknown)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "telecom_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telecom Phone"
          },
          "telecom_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telecom Email"
          }
        },
        "type": "object",
        "title": "PatientUpdate",
        "description": "Schema for updating a patient (partial)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}